Update Covariant Script Deployer/Form1.cs

+++ Covariant Script Deployer/Form1.cs
@@ -72,7 +72,7 @@ namespace Covariant_Script_Deployer
                 string path = info.Split('@')[0].Replace('/', '\\');
                 if (path.Length != 0)
                 {
-                    label6.Text = "正在部署..." + (progress / dist_info.Length) * 100 + "%";
+                    label6.Text = "正在部署..." + (int)(progress / dist_info.Length) * 100 + "%";
                     Application.DoEvents();
                     path = textBox2.Text + "\\" + path;
                     File.WriteAllText(path + ".md5", GetMD5HashFromFile(path));
